+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added support for `mcp check` command to show list of installed agents on the user PC
 - Added support for `mcp rm` command to remove mcp servers
 - Added support for `mcp list` command to show configured mcp servers